requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.5
Flask-Babel==4.0.0
Flask-Compress==1.14
Flask-Talisman==1.1.0
Flask-Blueprint==1.3.0

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2

redis==5.0.1
celery==5.3.4

prometheus-client==0.19.0
sentry-sdk[flask]==1.38.0
elastic-apm[flask]==6.19.0

pandas==2.1.4
numpy==1.26.2

python-dateutil==2.8.2
click==8.1.7
eventlet==0.33.3
gunicorn==21.2.0

Flask-Mail==0.9.1

# Reports & Export
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9
Pillow==10.1.0

prometheus-flask-exporter==0.23.2